[reorg] Move MultipleSubst into new layout.
