+++ src/meson.build
@@ -93,9 +93,12 @@ hb_base_sources = files(
   'hb-ot-layout-gpos-table.hh',
   'hb-ot-layout-gsub-table.hh',
   'OT/Layout/GSUB/Common.hh',
+  'OT/Layout/GSUB/Sequence.hh',
   'OT/Layout/GSUB/SingleSubstFormat1.hh',
   'OT/Layout/GSUB/SingleSubstFormat2.hh',
   'OT/Layout/GSUB/SingleSubst.hh',
+  'OT/Layout/GSUB/MultipleSubstFormat1.hh',
+  'OT/Layout/GSUB/MultipleSubst.hh',
   'hb-ot-layout-gsubgpos.hh',
   'hb-ot-layout-jstf-table.hh',
   'hb-ot-layout.cc',
